refactor(examples): inline CLI formatting helpers and trim hermes README
Remove the shared examples/hermes/_fmt.py module by inlining the helpers it
provided into the two scripts that used it, so each script is self-contained.
Also condense the README prose and use a placeholder model path.

Key changes:
- Inline formatting helpers into start_session.py and set_reward.py
- Delete examples/hermes/_fmt.py and its README entry
- Simplify quick-start prose; use actor.path=/path/to/your_model
- Switch config.yaml actor backend to megatron:d1
